requirements.txt
python-telegram-bot==20.5
nest_asyncio